refactor: Files structure
